+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 2.0.0
+
+- New class `SheetMap` that allows you to store data as a map
+- New class `SheetInteractionHandler` for more precise control of a sheet
+- Renamed `Table` to `SheetColumn` as the old name wasn't as clear and also interfered with the Flutter widget `Table` causing import inconveniences
+- Fixed `SheetColumn` encoding input data twice when writing to a sheet
+- Changed package description (Previous one was incomplete)
+- Added the ability to delete/clear cells
+
+
 ## 1.0.0
 
 - Initial version.